Move header into fcc_bridge sub folder
