+++ include/fcc_bridge/errors.hpp
