+++ src/fcc_bridge_node_ros.cpp
@@ -11,7 +11,7 @@
 #include <common_package/topic_names.hpp>
 
 // FCC Bridge header
-#include "fcc_bridge_node.hpp"
+#include "fcc_bridge/fcc_bridge_node.hpp"
 
 namespace fcc_bridge {
