+++ src/fcc_bridge_node_safety.cpp
@@ -6,7 +6,7 @@
 #include <set>
 
 // FCC Bridge header
-#include "fcc_bridge_node.hpp"
+#include "fcc_bridge/fcc_bridge_node.hpp"
 
 namespace fcc_bridge {
