+++ src/fcc_bridge_node_telemetry.cpp
@@ -3,7 +3,7 @@
 //
 
 // FCC Bridge header
-#include "fcc_bridge_node.hpp"
+#include "fcc_bridge/fcc_bridge_node.hpp"
 
 namespace fcc_bridge {
