+++ test/include/fcc_bridge_node_wrapper.hpp
@@ -7,7 +7,7 @@
 
 #include <gtest/gtest.h>
 
-#include "fcc_bridge_node.hpp"
+#include "fcc_bridge/fcc_bridge_node.hpp"
 
 namespace fcc_bridge::test {
